A MAJOR CHANGE
so many changes. 

Now the program adds Scenarios for testing when there are no scenarios (Similar to Instructor and Student Users for login testing). the SimulationGUI loads scenario from the Selected Scenario from the ScenarioSelectedGUI.

When the student logs in, you have access to that logged in student by calling SimulationManager.CURRENT_USER, Very useful when recording a Completed scenario after the student finishes the simulation

there are few more upgrades

the timer is fully functional in terms of setting timer

test the Scenario with the name of patient as Eric first. you might laugh :)