Updated GUIs on how they look and is extremely Debugged
New Button in Proffesor GUI (Print all student Records) which suppose to print each student record from a class. (Not implemented)

The program is basically completed exept the printing. 
(There are some implementation as a guide throughout the program located
* ProfessorGUI :   located in the print this Student Record's  actionEvent. It prints the table of all the completed scenarios that the selected student made. (line 1388)
*SimulationGUI : located in its print method. it is the print button on the GUI
)
They work perfectly so they are extremely useful


You can use these examples to complete the printing functionality of this program. In an email I will give you the TEMPLATES to use as a guide when implementing. You might need to work/implement with the Printer classes, SimulationGUI, CompletedScenario to make it work.