Extra stuff on the side of this project and NOT MANDATORY: The Student practice version of the program Development.
There are 3 new java source files that are essentially just copies from the sources From some of the GUIs and SimulationManager.java. They are all indentifiable with its part of its name "Practice".  

The Student practice version basically should start with scenario Selection. In the selection, the student is asked to load from a file, which is basically the scenario file that contains a list of scenarios. the loading can be implemented by calling the DataIO' s method called loadScenarios. Then the student is able to chouse from the list and then practice. In the practice version, the problems are unlimited time. When the student is done with the practice, he or she goes back to Selection.

Remember, this is not mandatory but you are welcome to work  on it. It should be way easier to implement since we have most of the code already; we just need to do few tweaks and it is all set.

for the student version, only you can work with the java files that contains in its name "Practice" , so we dont mess up the original Project we are completing.