Awsome News Everyone:
The PROGRAM IS FULLY OPERATIONAL !!!!!

There might be bugs out there but as far as I debug it, I have not found any yet.

The source code might need some cleanup and documentation for easier future reference and debugging.

there might be typos.

besides all these issue, I believe this program is ready for release. 
This is a RELEASE CANDIDATE (RC) of eMAR 
:)